Implemented Recipe changes

+++ src/main/java/com/bizzjen/cipes/repository/GroceryRepository.java
@@ -3,7 +3,5 @@ package com.bizzjen.cipes.repository;
 import com.bizzjen.cipes.entity.Grocery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface GroceryRepository extends JpaRepository<Grocery, BigDecimal> {
+public interface GroceryRepository extends JpaRepository<Grocery, Long> {
 }
