+++ src/main/java/com/bizzjen/cipes/repository/KitchenItemRepository.java
@@ -3,7 +3,5 @@ package com.bizzjen.cipes.repository;
 import com.bizzjen.cipes.entity.KitchenItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface KitchenItemRepository extends JpaRepository<KitchenItem, BigDecimal> {
+public interface KitchenItemRepository extends JpaRepository<KitchenItem, Long> {
 }
